+++ spec/dummy/config/initializers/gds_api.rb
@@ -1,0 +1,1 @@
+require 'gds_api/base'